Refactor configurable parameters
