+++ external/anomaly/anomaly_segmentation/configs/stfpm/configuration.py
@@ -1,5 +1,5 @@
 """
-Configurable parameters for STFPM anomaly classification task
+Configurable parameters for STFPM anomaly segmentation task
 """
 
 # Copyright (C) 2021 Intel Corporation
@@ -17,95 +17,11 @@ Configurable parameters for STFPM anomaly classification task
 # and limitations under the License.
 
 from attr import attrs
-from ote_anomalib.configs.configuration import BaseAnomalyConfig
-from ote_anomalib.configs.configuration_enums import EarlyStoppingMetrics, ModelBackbone
-from ote_sdk.configuration.elements import (
-    ParameterGroup,
-    add_parameter_group,
-    configurable_float,
-    configurable_integer,
-    selectable,
-    string_attribute,
-)
-from ote_sdk.configuration.model_lifecycle import ModelLifecycle
+from ote_anomalib.configs.stfpm import STFPMAnomalyBaseConfig
 
 
 @attrs
-class STFPMAnomalySegmentationConfig(BaseAnomalyConfig):
+class STFPMAnomalySegmentationConfig(STFPMAnomalyBaseConfig):
     """
-    Configurable parameters for STFPM anomaly classification task.
+    Configurable parameters for STFPM anomaly segmentation task.
     """
-
-    header = string_attribute("Configuration for STFPM")
-    description = header
-
-    @attrs
-    class ModelParameters(ParameterGroup):
-        """
-        Parameter Group for training model
-        """
-
-        header = string_attribute("Model Parameters")
-        description = header
-
-        backbone = selectable(
-            default_value=ModelBackbone.RESNET18,
-            header="Model Backbone",
-            description="Pre-trained backbone used for teacher and student network",
-        )
-
-        lr = configurable_float(
-            default_value=0.4,
-            header="Learning Rate",
-            min_value=1e-3,
-            max_value=1,
-            description="Learning rate used for optimizing the Student network.",
-        )
-
-        momentum = configurable_float(
-            default_value=0.9,
-            header="Momentum",
-            min_value=0.1,
-            max_value=1.0,
-            description="Momentum used for SGD optimizer",
-        )
-
-        weight_decay = configurable_float(
-            default_value=0.0001,
-            header="Weight Decay",
-            min_value=1e-5,
-            max_value=1,
-            description="Decay for SGD optimizer",
-        )
-
-        @attrs
-        class EarlyStoppingParameters(ParameterGroup):
-            """
-            Early stopping parameters
-            """
-
-            header = string_attribute("Early Stopping Parameters")
-            description = header
-
-            metric = selectable(
-                default_value=EarlyStoppingMetrics.IMAGE_F1,
-                header="Early Stopping Metric",
-                description="The metric used to determine if the model should stop training",
-            )
-
-            patience = configurable_integer(
-                default_value=10,
-                min_value=1,
-                max_value=100,
-                header="Early Stopping Patience",
-                description="Number of epochs to wait for an improvement in the monitored metric. If the metric has "
-                "not improved for this many epochs, the training will stop and the best model will be "
-                "returned.",
-                warning="Setting this value too low might lead to underfitting. Setting the value too high will "
-                "increase the training time and might lead to overfitting.",
-                affects_outcome_of=ModelLifecycle.TRAINING,
-            )
-
-        early_stopping = add_parameter_group(EarlyStoppingParameters)
-
-    model = add_parameter_group(ModelParameters)
